pyproject.toml
[tool.poetry]
name = "momo"
version = "0.1.0"
description = ""
authors = ["sliverKi <lovee2756@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
